Move Mode from AppData to Ball data and reset vel
For ease of use, we add the ball mode to the Ball struct itself. Not
sure this is strictly correct as the mode of the ball is more a function
of the app than the ball itself perhaps, but it certainly makes it
easier to have the Ball around when changing the mode so we can also
reset the ball's velocity when it is released.